+++ iqm_qaoa/src/iqm/applications/maxcut.py
@@ -34,11 +34,11 @@ Example:
         from iqm.applications.maxcut import MaxCutInstance
         from iqm.applications.maxcut import maxcut_generator
         from iqm.applications.maxcut import goemans_williamson
-        from iqm.applications.maxcut import greedy_maxcut
+        from iqm.applications.maxcut import greedy_max_cut
 
         for my_instance in maxcut_generator(graph_size, n_instances):  # Generates problem instances.
             gw_solution = goemans_williamson(my_instance)  # Solution bitstring obtained from GW.
-            greedy_solution = greedy_maxcut(my_instance)  # Solution bitstring obtained from greedy.
+            greedy_solution = greedy_max_cut(my_instance)  # Solution bitstring obtained from greedy.
             my_instance.cut_size(gw_solution)  # Check the GW solution cut size.
 
         my_maxcut_instance = MaxCutInstance(my_graph)  # Problem instance from a graph.
@@ -47,16 +47,22 @@ Example:
 
 from collections.abc import Iterator
 import random
+from typing import Literal
 
 import cvxpy as cp
 from dimod import BinaryQuadraticModel
-from iqm.applications.qubo import QUBOInstance, relabel_graph_nodes
+from iqm.applications.graph_utils import (
+    EDGE_ATTR_PRIORITY,
+    _generate_desired_graph,
+    _get_attr_with_priority,
+    relabel_graph_nodes,
+)
+from iqm.applications.qubo import QUBOInstance
 import networkx as nx
 import numpy as np
 from scipy.linalg import eigh
 
 
-# pylint: disable=anomalous-backslash-in-string
 class MaxCutInstance(QUBOInstance):
     r"""The maxcut instance class.
 
@@ -140,7 +146,6 @@ class MaxCutInstance(QUBOInstance):
         return cut
 
 
-# pylint: disable=anomalous-backslash-in-string
 class WeightedMaxCutInstance(QUBOInstance):
     r"""The weighted maxcut instance class.
 
@@ -161,27 +166,34 @@ class WeightedMaxCutInstance(QUBOInstance):
     """
 
     def __init__(self, graph: nx.Graph, break_z2: bool = False) -> None:
-        if set(graph.nodes) != set(range(len(graph))):
-            raise ValueError("Graph nodes should be labelled by integers starting with 0.")
-        for n1, n2, data in graph.edges(data=True):
-            if "weight" not in data:
-                raise ValueError(f"The edge between nodes {n1} and {n2} is missing the 'weight' attribute.")
-            if not isinstance(data["weight"], float | int):
-                raise TypeError(
-                    f"The edge between nodes {n1} and {n2} has a 'weight' of type {type(data['weight']).__name__}, "
-                    f"expected ``float`` or ``int``."
+        self._graph, self.orig_to_new_labels, self.new_to_orig_labels = relabel_graph_nodes(graph)
+
+        qubo_mat = np.zeros((self._graph.number_of_nodes(), self._graph.number_of_nodes()))
+
+        for n1, n2, data in self._graph.edges(data=True):
+            value = _get_attr_with_priority(data, EDGE_ATTR_PRIORITY)
+
+            if value is None:
+                raise ValueError(
+                    f"The edge between nodes {self.new_to_orig_labels[n1]} and {self.new_to_orig_labels[n2]} is missing"
+                    f" one of the required attributes ({', '.join(EDGE_ATTR_PRIORITY)})."
                 )
 
-        self._graph = graph
-        self._break_z2 = break_z2
+            if not isinstance(value, (float, int)):
+                raise TypeError(
+                    f"The edge between nodes {self.new_to_orig_labels[n1]} and {self.new_to_orig_labels[n2]} has a "
+                    f"value of type {type(value).__name__}, expected ``float`` or ``int``."
+                )
 
-        qubo_mat = np.zeros((self._graph.number_of_nodes(), self._graph.number_of_nodes()), dtype=int)
-        for i, j, edge_data in self._graph.edges(data=True):
-            qubo_mat[i, j] += 2 * edge_data["weight"]
-            qubo_mat[i, i] += -edge_data["weight"]
-            qubo_mat[j, j] += -edge_data["weight"]
+            qubo_mat[n1, n2] += 2 * value
+            qubo_mat[n1, n1] += -value
+            qubo_mat[n2, n2] += -value
+
         bqm = BinaryQuadraticModel(qubo_mat, vartype="BINARY")
+
         super().__init__(bqm)
+
+        self._break_z2 = break_z2
         if self._break_z2:
             node_to_fix = max(bqm.variables, key=bqm.degree)
             self.fix_variables({node_to_fix: 1})
@@ -232,17 +244,17 @@ class WeightedMaxCutInstance(QUBOInstance):
         return cut
 
 
-# pylint: disable=anomalous-backslash-in-string
-def maxcut_generator(
+def maxcut_generator(  # noqa: PLR0913
     n: int,
     n_instances: int,
     *,
-    graph_family: str = "erdos-renyi",
+    graph_family: Literal["regular", "erdos-renyi"] = "erdos-renyi",
     p: float = 0.5,
     d: int = 3,
     break_z2: bool = False,
-    seed: int = 1337,
+    seed: int | None = None,
     enforce_connected: bool = False,
+    max_iterations: int = 1000,
 ) -> Iterator[MaxCutInstance]:
     r"""The generator function for generating random maxcut problem instances.
 
@@ -262,27 +274,18 @@ def maxcut_generator(
         break_z2: Optional bool indicating whether the :math:`\mathbb{Z}_2` symmetry should be explicitly broken
             in the problem instances.
         seed: Optional random seed for generating the problem instances.
-        enforce_connected: A bool stating whether it is required that the random graphs are connected.
+        enforce_connected: True iff it is required that the random graphs are connected.
+        max_iterations: In case ``enforce_connected`` is ``True``, the function generates random graphs in a ``while``
+            loop until it finds a connected one. If it doesn't find a connected one after ``max_iterations``, it raises
+            an error.
 
     Yields:
         Problem instances of :class:`MaxCutInstance` randomly constructed in accordance to the input parameters.
 
     """
-
-    def _generate_desired_graph(graph_family: str, n: int, p: float, d: int, seed: int) -> nx.Graph:
-        """Wrapper helper function to encapsulate the graph generation logic."""
-        if graph_family == "erdos-renyi":
-            return nx.erdos_renyi_graph(n, p, seed)
-        if graph_family == "regular":
-            return nx.random_regular_graph(d, n, seed)
-        raise ValueError("Invalid random graph type. Choose either 'regular' or 'erdos-renyi'.")
-
     for _ in range(n_instances):
-        while True:
-            g = _generate_desired_graph(graph_family, n, p, d, seed)
-            seed = seed + 1  # Increment ``seed``, so that we don't keep generating the same graph over and over again.
-            if nx.is_connected(g) or not enforce_connected:
-                break
+        g = _generate_desired_graph(graph_family, n, p, d, seed, enforce_connected, max_iterations)
+
         yield MaxCutInstance(g, break_z2=break_z2)
 
 
@@ -328,7 +331,6 @@ def greedy_max_cut(max_cut_problem: MaxCutInstance | nx.Graph) -> str:
     return "".join(current_solution)
 
 
-# pylint: disable=too-many-locals
 def goemans_williamson(max_cut_problem: MaxCutInstance | nx.Graph) -> str:
     """Runs the Goemans-Williamson algorithm for maxcut, returning a solution bitstring.